backend/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
openpyxl
python-multipart
scikit-learn
plotly
matplotlib
google-generativeai
python-dotenv
